The cast to int isn't needed for python3.

+++ homeassistant/components/light/xiaomi_miio.py
@@ -205,7 +205,7 @@ class XiaomiPhilipsGenericLight(Light):
         """Turn the light on."""
         if ATTR_BRIGHTNESS in kwargs:
             brightness = kwargs[ATTR_BRIGHTNESS]
-            percent_brightness = int(ceil(100 * brightness / 255.0))
+            percent_brightness = ceil(100 * brightness / 255.0)
 
             _LOGGER.debug(
                 "Setting brightness: %s %s%%",
@@ -236,7 +236,7 @@ class XiaomiPhilipsGenericLight(Light):
             _LOGGER.debug("Got new state: %s", state)
 
             self._state = state.is_on
-            self._brightness = int(ceil((255/100.0) * state.brightness))
+            self._brightness = ceil((255/100.0) * state.brightness)
 
         except DeviceException as ex:
             _LOGGER.error("Got exception while fetching the state: %s", ex)
@@ -307,7 +307,7 @@ class XiaomiPhilipsLightBall(XiaomiPhilipsGenericLight, Light):
 
         if ATTR_BRIGHTNESS in kwargs:
             brightness = kwargs[ATTR_BRIGHTNESS]
-            percent_brightness = int(ceil(100 * brightness / 255.0))
+            percent_brightness = ceil(100 * brightness / 255.0)
 
             _LOGGER.debug(
                 "Setting brightness: %s %s%%",
@@ -332,7 +332,7 @@ class XiaomiPhilipsLightBall(XiaomiPhilipsGenericLight, Light):
             _LOGGER.debug("Got new state: %s", state)
 
             self._state = state.is_on
-            self._brightness = int(ceil((255/100.0) * state.brightness))
+            self._brightness = ceil((255/100.0) * state.brightness)
             self._color_temp = self.translate(
                 state.color_temperature,
                 CCT_MIN, CCT_MAX,
